backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]>=0.29.0
motor>=3.3.2
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.3.4
python-dotenv>=1.0.1
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.1.1
pymupdf>=1.24.0
groq>=0.9.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
